refactor(kanban): split the board component into focused modules
